refactor: home페이지 코드 분리 및 리팩토링
