+++ index.html
@@ -3,9 +3,12 @@
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/public/favicon.svg" />
-    <script type="text/javascript" src="//dapi.kakao.com/v2/maps/sdk.js?appkey=%VITE_KAKAO_KEY%"></script>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>RIDERS</title>
+    <script
+      type="text/javascript"
+      src="//dapi.kakao.com/v2/maps/sdk.js?appkey=%VITE_KAKAO_KEY%&libraries=services,clusterer"
+    ></script>
   </head>
   <body>
     <div id="root"></div>
