+++ src/axiosInstance/information.js
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+const informationInstance = axios.create({
+  baseURL: import.meta.env.VITE_PATH_INFORMATION_URL
+});
+
+export default informationInstance;
