+++ src/constants/homeConstants.js
@@ -1,0 +1,1 @@
+export const HOME_BUTTONS_LIST = ["전체", "자전거길", "화장실", "공기주입기", "인증센터", "급수대"];
