+++ src/pages/home/Home.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useRef } from "react";
+import { useState, useEffect } from "react";
 import styled from "styled-components";
 import Filter from "./Filter";
 import { getFilterRoad } from "../../api/FilterRoadInformation";
@@ -10,6 +10,7 @@ import useInsertFeed from "../../hooks/useInsertFeed";
 import Swal from "sweetalert2";
 import { useNavigate } from "react-router-dom";
 import useUserStore from "../../store/useUserStore";
+import calculateTotalDistance from "../../utils/calculateTotalDistance";
 
 const Home = () => {
   const [filterData, setFilterData] = useState([]);
@@ -24,20 +25,10 @@ const Home = () => {
   const { mutate: insertFeed, isLoading, isError } = useInsertFeed();
 
   const navigate = useNavigate();
-  // console.log(mapCenter);
-  // console.log(filterData);
   //현위치
   const location = useGeoLocation();
   const [positions, setPositions] = useState([]);
   const [amenityDatas, setAmenityDatas] = useState([]);
-  // console.log("amenityDatas =>", amenityDatas);
-
-  // console.log(filterData);
-  //현위치
-  // useEffect(() => {
-  //   setCenterCoord(location.center);
-  // }, []);
-  // console.log(location);
 
   useEffect(() => {
     const getAllRoadData = async () => {
@@ -53,42 +44,9 @@ const Home = () => {
   }, []);
 
   useEffect(() => {
-    // console.log(polyline);
     if (polyline?.roadLine?.length > 0) {
-      const haversineDistance = (coords1, coords2) => {
-        const toRad = (value) => (value * Math.PI) / 180;
-        const R = 6371; // 지구의 반경(km)
-        const dLat = toRad(coords2.lat - coords1.lat);
-        const dLng = toRad(coords2.lng - coords1.lng);
-
-        const a =
-          Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-          Math.cos(toRad(coords1.lat)) * Math.cos(toRad(coords2.lat)) * Math.sin(dLng / 2) * Math.sin(dLng / 2);
-
-        const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-        return R * c; // 두 좌표 사이의 거리(km)
-      };
-
-      const calculateTotalDistance = (roadLine) => {
-        let totalDistance = 0;
-        for (let i = 0; i < roadLine.length - 1; i++) {
-          const start = { lat: roadLine[i].LINE_XP, lng: roadLine[i].LINE_YP };
-          const end = { lat: roadLine[i + 1].LINE_XP, lng: roadLine[i + 1].LINE_YP };
-          totalDistance += haversineDistance(start, end);
-        }
-        return totalDistance.toFixed(2); // 총 거리(km)를 소수점 두 자리로 반환
-      };
-
       const totalDistance = calculateTotalDistance(polyline.roadLine);
       setTotalDistance(totalDistance);
-      // console.log("총 거리:", totalDistance, "km");
-    }
-  }, [polyline]);
-
-  // 버튼 클릭 시 현재 위치로 지도 중심 이동
-  // polyline의 첫 번째 좌표가 변경되면 mapCenter를 업데이트
-  useEffect(() => {
-    if (polyline?.roadLine?.length > 0) {
       setMapCenter({ lat: polyline.roadLine[0].LINE_XP, lng: polyline.roadLine[0].LINE_YP });
     }
   }, [polyline]);
